+++ Gems/Atom/RHI/Code/Source/RHI/CpuProfilerImpl.cpp
@@ -382,5 +382,64 @@ namespace AZ
                 m_cachedTimeRegionMutex.unlock();
             }
         }
-    }
-}
+
+        // --- CpuProfilingStatisticsSerializer ---
+
+        CpuProfilingStatisticsSerializer::CpuProfilingStatisticsSerializer(const AZStd::ring_buffer<RHI::CpuProfiler::TimeRegionMap>& continuousData)
+        {
+            // Create serializable entries
+            for (const auto& timeRegionMap : continuousData)
+            {
+                for (const auto& threadEntry : timeRegionMap)
+                {
+                    for (const auto& cachedRegionEntry : threadEntry.second)
+                    {
+                        m_cpuProfilingStatisticsSerializerEntries.insert(
+                            m_cpuProfilingStatisticsSerializerEntries.end(),
+                            cachedRegionEntry.second.begin(),
+                            cachedRegionEntry.second.end());
+                    }
+                }
+            }
+        }
+
+        void CpuProfilingStatisticsSerializer::Reflect(AZ::ReflectContext* context)
+        {
+            if (auto* serializeContext = azrtti_cast<AZ::SerializeContext*>(context))
+            {
+                serializeContext->Class<CpuProfilingStatisticsSerializer>()
+                    ->Version(1)
+                    ->Field("cpuProfilingStatisticsSerializerEntries", &CpuProfilingStatisticsSerializer::m_cpuProfilingStatisticsSerializerEntries)
+                    ;
+            }
+
+            CpuProfilingStatisticsSerializerEntry::Reflect(context);
+        }
+
+        // --- CpuProfilingStatisticsSerializerEntry ---
+
+        CpuProfilingStatisticsSerializer::CpuProfilingStatisticsSerializerEntry::CpuProfilingStatisticsSerializerEntry(const RHI::CachedTimeRegion& cachedTimeRegion)
+        {
+            m_groupName = cachedTimeRegion.m_groupRegionName->m_groupName;
+            m_regionName = cachedTimeRegion.m_groupRegionName->m_regionName;
+            m_stackDepth = cachedTimeRegion.m_stackDepth;
+            m_startTick = cachedTimeRegion.m_startTick;
+            m_endTick = cachedTimeRegion.m_endTick;
+        }
+
+        void CpuProfilingStatisticsSerializer::CpuProfilingStatisticsSerializerEntry::Reflect(AZ::ReflectContext* context)
+        {
+            if (auto* serializeContext = azrtti_cast<AZ::SerializeContext*>(context))
+            {
+                serializeContext->Class<CpuProfilingStatisticsSerializerEntry>()
+                    ->Version(1)
+                    ->Field("groupName", &CpuProfilingStatisticsSerializerEntry::m_groupName)
+                    ->Field("regionName", &CpuProfilingStatisticsSerializerEntry::m_regionName)
+                    ->Field("stackDepth", &CpuProfilingStatisticsSerializerEntry::m_stackDepth)
+                    ->Field("startTick", &CpuProfilingStatisticsSerializerEntry::m_startTick)
+                    ->Field("endTick", &CpuProfilingStatisticsSerializerEntry::m_endTick)
+                    ;
+            }
+        }
+    } // namespace RHI
+} // namespace AZ
